+++ src/app.rs
@@ -217,11 +217,13 @@ where
 				None => Vec::new(),
 			}
 		} else {
-			recorder.record_sync(if self.settings.record.flag.keys.is_some() {
-				self.settings.input_state
-			} else {
-				None
-			})
+			recorder.record_sync(
+				if self.settings.record.flag.action_keys.is_some() {
+					self.settings.input_state
+				} else {
+					None
+				},
+			)
 		}
 	}