+++ src/util/keys.rs
@@ -2,11 +2,21 @@ use device_query::Keycode;
 use std::fmt;
 use std::str::FromStr;
 
+/* Types of key bindings. */
+#[derive(Debug)]
+pub enum KeyType {
+	ActionKeys,
+	CancelKeys,
+}
+
 /* Operational keys and combinations */
 #[derive(Debug)]
 pub struct ActionKeys {
 	key_groups: Vec<Vec<Keycode>>,
 }
+
+/* Alias for cancel keys */
+pub type CancelKeys = ActionKeys;
 
 /* Display implementation for user-facing output */
 impl fmt::Display for ActionKeys {
@@ -27,18 +37,6 @@ impl fmt::Display for ActionKeys {
 	}
 }
 
-/* Default initialization values for ActionKeys */
-impl Default for ActionKeys {
-	fn default() -> Self {
-		Self {
-			key_groups: vec![
-				vec![Keycode::LAlt, Keycode::S],
-				vec![Keycode::LAlt, Keycode::Enter],
-			],
-		}
-	}
-}
-
 impl ActionKeys {
 	/**
 	 * Create a new ActionKeys object.
@@ -48,6 +46,29 @@ impl ActionKeys {
 	 */
 	pub fn new(key_groups: Vec<Vec<Keycode>>) -> Self {
 		Self { key_groups }
+	}
+
+	/**
+	 * Return the default keys based on the given key type.
+	 *
+	 * @param  key_type
+	 * @return ActionKeys
+	 */
+	pub fn default(key_type: KeyType) -> Self {
+		match key_type {
+			KeyType::ActionKeys => Self {
+				key_groups: vec![
+					vec![Keycode::LAlt, Keycode::S],
+					vec![Keycode::LAlt, Keycode::Enter],
+				],
+			},
+			KeyType::CancelKeys => Self {
+				key_groups: vec![
+					vec![Keycode::LControl, Keycode::D],
+					vec![Keycode::Escape],
+				],
+			},
+		}
 	}
 
 	/**
@@ -66,9 +87,10 @@ impl ActionKeys {
 	 * Parse ActionKeys from a string.
 	 *
 	 * @param  keys
+	 * @param  key_type
 	 * @return ActionKeys
 	 */
-	pub fn parse(keys: &str) -> Self {
+	pub fn parse(keys: &str, key_type: KeyType) -> Self {
 		let keys = keys
 			.split(',')
 			.filter_map(|keys| {
@@ -80,7 +102,7 @@ impl ActionKeys {
 			})
 			.collect::<Vec<Vec<Keycode>>>();
 		if keys.is_empty() {
-			Self::default()
+			Self::default(key_type)
 		} else {
 			Self::new(keys)
 		}
@@ -112,7 +134,7 @@ mod tests {
 	#[test]
 	fn test_action_keys() {
 		let keys_str = "LControl-Q,LControl-W";
-		let keys = ActionKeys::parse(keys_str);
+		let keys = ActionKeys::parse(keys_str, KeyType::ActionKeys);
 		assert_eq!(keys_str, keys.to_string());
 		assert_eq!(
 			vec![
@@ -127,26 +149,30 @@ mod tests {
 		assert!(!keys.check(vec![Keycode::W]));
 		assert!(keys.check(vec![Keycode::LControl, Keycode::Q]));
 		assert!(keys.check(vec![Keycode::LControl, Keycode::W]));
-		assert!(!ActionKeys::parse("S").check(vec![Keycode::S, Keycode::Slash]));
-		assert!(ActionKeys::parse("X,Y").check(vec![Keycode::X]));
-		assert!(ActionKeys::parse("LControl-J,A,B,C")
+		assert!(!ActionKeys::parse("S", KeyType::ActionKeys)
+			.check(vec![Keycode::S, Keycode::Slash]));
+		assert!(
+			ActionKeys::parse("X,Y", KeyType::ActionKeys).check(vec![Keycode::X])
+		);
+		assert!(ActionKeys::parse("LControl-J,A,B,C", KeyType::ActionKeys)
 			.check(vec![Keycode::LControl, Keycode::J]));
-		assert!(ActionKeys::parse("LControl-A,X,Y").check(vec![Keycode::Y]));
+		assert!(ActionKeys::parse("LControl-A,X,Y", KeyType::ActionKeys)
+			.check(vec![Keycode::Y]));
 		assert_eq!(
-			ActionKeys::default().key_groups,
-			ActionKeys::parse("LCxntrxl-WW").key_groups
+			ActionKeys::default(KeyType::CancelKeys).key_groups,
+			ActionKeys::parse("LCxntrxl-WW", KeyType::CancelKeys).key_groups
 		);
 		assert_eq!(
 			vec![vec![Keycode::X]],
-			ActionKeys::parse("test,X,...").key_groups
+			ActionKeys::parse("test,X,...", KeyType::ActionKeys).key_groups
 		);
 		assert_eq!(
 			vec![vec![Keycode::LControl], vec![Keycode::X]],
-			ActionKeys::parse("LControl-test,X").key_groups
+			ActionKeys::parse("LControl-test,X", KeyType::ActionKeys).key_groups
 		);
 		assert_eq!(
 			vec![&Keycode::A, &Keycode::C],
-			ActionKeys::parse("A-B,C-D,...").get_primary()
+			ActionKeys::parse("A-B,C-D,...", KeyType::ActionKeys).get_primary()
 		);
 	}
 }